Fix bug with __str__ in python 3
The __str__ methods of Settings and Subscription previously returned
bytes in python 3 and caused errors because the str function expects
(unicode) strings in python 3.

This commit fixes the bug. It also works for non-ascii
translation strings in python 2 as mentioned in b1e0b23cd7b4e4e.

The commit also fixes problems when the username of a user has non-ascii
characters.